+++ Sources/VenmoPayments/VenmoClient.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public class VenmoClient {
+    
+    // TODO: to be implemented in a future PR
+}