+++ UnitTests/VenmoPaymentsTests/VenmoClient_Tests.swift
@@ -1,0 +1,6 @@
+import XCTest
+
+public class VenmoClient_Tests: XCTestCase {
+    
+    // TODO: to be implemented in a future PR
+}